+++ libraries/transformer/src/main/java/androidx/media3/transformer/ResumedExportOperation.java
@@ -527,11 +527,14 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
     }
 
     @Override
-    public void onEnded(long approximateDurationMs, long fileSizeBytes) {
-      exportResultBuilder
-          .setApproximateDurationMs(approximateDurationMs)
-          .setFileSizeBytes(fileSizeBytes);
+    public void onEnded(long approximateDurationMs) {
+      exportResultBuilder.setApproximateDurationMs(approximateDurationMs);
       checkNotNull(transformerInternal).endWithCompletion();
+    }
+
+    @Override
+    public void onFileSizeBytesAvailable(long fileSizeBytes) {
+      exportResultBuilder.setFileSizeBytes(fileSizeBytes);
     }
 
     @Override