+++ libraries/transformer/src/test/java/androidx/media3/transformer/MuxerWrapperTest.java
@@ -722,7 +722,10 @@ public class MuxerWrapperTest {
     public void onSampleWrittenOrDropped() {}
 
     @Override
-    public void onEnded(long approximateDurationMs, long fileSizeBytes) {}
+    public void onEnded(long approximateDurationMs) {}
+
+    @Override
+    public void onFileSizeBytesAvailable(long fileSizeBytes) {}
 
     @Override
     public void onError(ExportException exportException) {}